textarea -> contenteditable div
These changes update the text areas for bullet input and remarks output so that both can be re-implemented using content editable divs.

This improves their overall look and feel, and also enables the possibility of conditional formatting in the future (i.e. highlighting acronyms/abbreviations that are being hovered over.